backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
openai==1.10.0
pinecone-client==3.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.26.0
PyPDF2==3.0.1
beautifulsoup4==4.12.3
redis==5.0.1
celery==5.3.6
boto3==1.34.34
